Finished compat changes for FP 10, assuming Ultimate tests succeed
- The version of CLIFp is now correctly displayed when using the -v
option instead of showing the target FP version twice

- Added read of flashpointPath when parsing the FP config (currently
unused but may be needed in future)

- Added read of jsonFolder path when parsing FP preferences so that the
services.json file can be found dynamically in case it moves in the
future

- Changed JSONServicesReader to take the host install path directly
instead of an Install instances pointer for resolving the fpPath macro.
This way the reader can be used when an Install instance is not
available (i.e. during checkInstallValidity)

- Updated target version strings and SHA256

- Further reordered mentions of Config and Services to be more
consistant with each other (config -> services)

- Heavily refactored checkInstallValidity to be more accurate for both
compat levels, fix an unreachable check of Config reading, and overall
be more through

- Noted that checkInstallValidity should be removed in favor of just
brute forcing the steps in the Install contstructor and checking after
construction if the instance is valid, if possible