Changed file vs dir detection rules for link -p switch
- The path ending with the .lnk extension now dictates that the path is
to a file, with all else being treated as a directory. This avoids the
annoyance of having to escape the trailing backslash in the previous
implementation due to Windows using \ as the double quote escape
character, since directories can now be passed without a trailing
backslash. Trailing backslashes will still work if properly escaped.