Implement inherited JSON structs/readers, move resolveFlashpointMacros
Moved resolveFlashpointMacros to FP::Install since the macro resolution
is directly tied to a given install; however, to make it idomatic it was
made a member function that references the instances paths. Therefore,
it can no longer be used without an FP::Install instance and as such
ServicesReader cannot be either. Because of this checkInstallValidity
will have to be removed in favor of the approach of simply trying to
instatiate an Install instance and then checking for validity after the
fact to know if it was valid.