Make task handling more asynchronous
Use signals/slots for task completion and a queued connection so that
after each task is completed (or during tasks that need to idle, i.e.
wait, blocking process) control returns to the Driver event loop.

This was its possible to for external signals to be received and
processed while the task queue is running, like a quit notice.