Downgrade zlib for QuaZip to v1.2.11 to avoid newer Qt symbol clash
Sometime after Qt 6.2.4 (I think about 6.3.0), Qt updated its built-in
zlib version from v1.2.11 to v1.2.12. Historically, Qt has always
prefixed its copy of zlib's symbols so that they don't clash with user
builds; however, v1.2.12 of zlib added `crc32_combine_gen()` and
crc32_combine_op()`, which for whatever reason Qt has not prefixed. As
such, if using that same zlib version in a user project, the definitions
for those symbols will clash between the Qt built-in and external zlib
builds.

For now this is worked around by dropping the external zlib version for
this project from v1.2.12 to v1.2.11, before those two symbols were
introduced. This is not a good long term solution. Hopefully in the
future Qt will prefix those symbols as well, or some other reasonable
means of using an external zlib separate from the built-in Qt one will
be possible.

Right now the only other way to do this is build Qt with the `-system-
zlib` option, but this requires an extra copy of Qt just for this case,
and potentially makes it impossible to do per-project builds of zlib
like is done here since Qt then needs to link to the same copy of zlib.
If Qt built like this is able to grab zlib from any location it is
available (i.e. via something like CMake's `find_dependency`) then it
shouldn't be too hard to build Qt using a static copy of zlib and when
using it in a project like this simply fetch zlib first before doing
`find_package(Qt6...)` and have it then use that copy of zlib (ensuring
they are the same version) when building said project.

If Qt instead hard codes the path to zlib.lib and its headers when being
built, it would basically be impossible to do per-project source builds
for zlib and they would have to prefix these new symbols.