Change task implementation to be fully polymorphic
This makes maintenance of each task much more straight forward and also
massively improves the ease with which this application can be ported to
other systems (i.e via a separate implementation .cpp for OS specific
implementation for each task). Additionally, the execution flow of
driver is now a lot more streamlined.

A caveat of this is that ErrorCode definitions are now somewhat of a
mess, being split between Core and Tasks, but still numbered as if only
belonging to Core. They likely should be reindexed to "belong" to each
task, just as with Commands.

Lastly this also required dropping support for the user entering
credentials in the case that Download tasks required them, though
realistically this should never be needed, unless FP starts requiring
user accounts, which is highly unlikely.

Originally the support was dropped because it would require an ugly
break of the polymorphic approach by checking explicitly for a Download
task and connecting to a special signal from that derived class as an
exception; however, instead Task could just have a "credentialsRequired"
signal that conceivably could be used for anything login related, which
would make it more reasonable to have as a base class signal.