Parent Deferred processes in TExec to Core
Can no longer set the parent of deferred processes to 'this' since they
are now created within TExec, as the TExec instance will die before the
process is finished (given its held by the static memebr
TExec::smActiveChildProcesses). So instead set their parent to the
parent of the TExec instance itself, which is Core.