Reset long task progress dialog instead of closing
Calling close() on the dialog causes the emission of its canceled()
signal, which was not the intended effect when closing it
programatically upon the completion of a long task.

Additionally, make the dialog a stack member variable since it doesn't
need to be deleted for any reason, given the new call to reset() is
adequate.