Use QApplication::arguments() instead of GetCommandLineW()
Originally the native windows version was used to grab the raw command
line parameters as the documentation for QApplication::arguments()
potentially implied that if the application was started in such a way
that its full path was part of the arguments list and that path had
spaces in it, then removing that path and the app's name would no longer
be as simple as discarding arguments[0] since that would just be the
path up until the first space.

However, this is not the case and even with spaces the entirety of the
first element in the arguments list is the path/app name, and therefore
they can easily be removed, dropping the need for separate handling via
GetCommandLineW().