Improve worker thread management
The motivation for this change is that when Driver finished, but window
are were still open and the call to QApplication::quit() is queued, if a
quit request is made by the user at this time (causing Controller to
emit quit) Driver is still be notified of the request and attempts to use
its closed log (the error for which isn't be seen as the application is
closed before the window is shown.

This is because when the quit request was implemented, the avoidance of
signaling the finished Driver instance was predicated on the fact that Qt
automatically disconnects connections when one of the objects involved
is destroyed, the thinking was that once Driver was finished, the "quit"
emission of Controller would be ignored since Driver was gone; however,
this didn't take into account that the current thread management
approach was setup so that Driver would only be deleted when the worker
thread finished, and the worker thread was only told to quit in
Controller's destructor, meaning that Driver wasn't deleted until the
application was already quitting.

This could have been worked around by adding a mFinished bool member to
Controller that is set once Driver finishes and is checked so that the
quit request is ignored if it's set, but it was then thought best to
cleanup Driver and stop the worker thread as soon as they were no longer
needed, instead of leaving them around until the end of main is reached.

Change worker thread management so that Driver is signaled to start when
the thread starts, and the thread is signaled to stop when Driver is
finished, with Controller simply starting the thread when `run()` is
called, instead of emitting an `operate()` signal. Additionally, Driver
schedules itself for deletion when finished, and Controller only
initiates `finisher()` once the thread has stopped.

Overall, the approach is now such that the worker thread and driver will
have completely finished (and driver will have been destroyed) before the
application is quit from user code.

This could be extended by making the worker thread a heap variable and
have it be deleted after its quit for additional ASAP cleanup, but
keeping QThread instance of a quit thread within Controller probably has
very little cost, so leaving its destruction until the application is
quit is fine.