Store process identifier as object name in DeferredProcessManager
This more tightly binds the identity of the process to the process
object itself, and also obviates the need to use a hash for
mManagedProcesses, which can now instead just be a set.

Additionally, this fixes an oversight in which processFinishedHandler
would remove a process from the management hash (therefore removing
access to its identity) and then pass that process to the stdout and
stderr buffer handlers, which would then try to retrieve its identity
from the same hash, leading to a null string being received.