Add support for 'flashpoint://' protocol
CLIFp can register itself to respond to use of the above URL scheme in
order to start titles. Additionally, the new 'share' command will
generate links based on titles that use the protocol and adds them to
the clipboard so that users can
more easily share them.

Also implements an https redirect page to be hosted on GitHub that
redirects to a FP protocol link. This is useful so that users can still
share links on messageing platforms that don't allow for custom scheme
hyperlinks directly.