Fix initial executable path determination
Was using QFileInfo::canonicalFilePath() et. al. before actual
filesystem resolution step, which meant that on Linux when forming a
path to something that needs to be swapped from its windows variant
(i.e. bat -> sh) a blank string would be created instead of the initial
path.

Use QDir::cleanPath() along with QFileInfo()::absoluteFilePath() to
obtain the same clean string but without the requirement of the path
actually existing.