Fix unintentionally shared help string buffer among Commands
This bug was never seen as only one Command is ever constructed per-run,
but for posterity we ensure that each Command instance has it's own help
string storage so that if for whatever reason in the future we make
multiple commands at once, the help string is not stuck as the one
created for the first command used.

In the long, once possible, the help string should be a constexpr member
that is made at compile time. We could do that now if using std::string,
but it's not a big deal.